把一些屎山清理了

+++ UserCode/Lib/usercode_vc/Coordinate_Trans.c
@@ -10,9 +10,9 @@ void CT_Refresh_abc(CT_data_t *CTdata,float a,float b,float c,float theta)
     = sqrt(2/3) * ((0)*a + (sqrt(3)/2)*b - (sqrt(3)/2)*c);
 
     CTdata->dq_value.d 
-    = (2/3) * ((cos(theta))*a + (cos(theta-2*PI/3))*b + (cos(theta+2*PI/3))*c);
+    = (2/3) * ((cosf(theta))*a + (cosf(theta-2*PI/3))*b + (cosf(theta+2*PI/3))*c);
      CTdata->dq_value.q 
-    = -(2/3) * ((sin(theta))*a + (sin(theta-2*PI/3))*b - (sin(theta+2*PI/3))*c);
+    = -(2/3) * ((sinf(theta))*a + (sinf(theta-2*PI/3))*b - (sinf(theta+2*PI/3))*c);
 }
 
 void CT_Refresh_AlphaBeta(CT_data_t *CTdata,float alpha,float beta,float theta)
@@ -25,24 +25,24 @@ void CT_Refresh_AlphaBeta(CT_data_t *CTdata,float alpha,float beta,float theta)
     = sqrt(2/3) * (-(1/2)*alpha - (sqrt(3)/2)*beta);
 
     CTdata->dq_value.d 
-    = (cos(theta))*alpha + (sin(theta))*beta;
+    = (cosf(theta))*alpha + (sinf(theta))*beta;
      CTdata->dq_value.q 
-    = -(sin(theta))*alpha + (cos(theta))*beta;
+    = -(sinf(theta))*alpha + (cosf(theta))*beta;
 }
 
 void CT_Refresh_dq(CT_data_t *CTdata,float d,float q,float theta)
 {
     CTdata->abc_value.a 
-    = cos(theta)*d - sin(theta)*q;
+    = cosf(theta)*d - sinf(theta)*q;
     CTdata->abc_value.b
-    = cos(theta-2*PI/3)*d - sin(theta-2*PI/3)*q;
+    = cosf(theta-2*PI/3)*d - sinf(theta-2*PI/3)*q;
     CTdata->abc_value.c
-   = cos(theta+2*PI/3)*d - sin(theta+2*PI/3)*q;
+   = cosf(theta+2*PI/3)*d - sinf(theta+2*PI/3)*q;
 
     CTdata->AlphaBeta_value.alpha 
-    = (cos(theta))*d - (sin(theta))*q;
+    = (cosf(theta))*d - (sinf(theta))*q;
      CTdata->AlphaBeta_value.beta 
-    = (sin(theta))*d + (cos(theta))*q;
+    = (sinf(theta))*d + (cosf(theta))*q;
 }
 
 void CT_get_AlphaBeta(CT_data_t *CTdata,float a,float b,float c)
